+++ src/components/Card.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import '../Card.css'
+import '../styles/Card.css'
 
 const Card = (props) => {
   return (
