+++ src/styles/MovieDetails.css
